+++ RELEASE_NOTES.md
@@ -9,3 +9,9 @@ Update input set object type to KBaseSets.ReadsAlignmentSet
 
 1.0.3:
 Add merge option to merge result gtf files
+
+1.0.5:
+Add a third run using merge reference annotation
+
+1.0.6:
+Splite StringTie app and StringTie app with novel isoform